+++ core/domain/question_services_test.py
@@ -14,7 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from core.domain import exp_domain
 from core.domain import question_domain
 from core.domain import question_services
 from core.domain import user_services
@@ -27,41 +26,6 @@ memcache_services = models.Registry.import_memcache_services()
 
 class QuestionServicesUnitTest(test_utils.GenericTestBase):
     """Test the question services module."""
-
-    def _create_valid_question_data(self, default_dest_state_name):
-        """Creates a valid question_data dict.
-
-        Args:
-            default_dest_state_name: str. The default destination state.
-
-        Returns:
-            dict. The default question_data dict.
-        """
-        state = exp_domain.State.create_default_state(
-            default_dest_state_name).to_dict()
-        solution = {
-            'answer_is_exclusive': False,
-            'correct_answer': 'Solution',
-            'explanation': {
-                'html': 'Solution explanation',
-                'audio_translations': {}
-            }
-        }
-        state['interaction']['id'] = 'TextInput'
-        state['interaction']['customization_args'] = {
-            'placeholder': 'Enter text here',
-            'rows': 1
-        }
-        state['interaction']['default_outcome']['labelled_as_correct'] = True
-        state['interaction']['default_outcome']['dest'] = None
-        state['interaction']['hints'].append({
-            'hint_content': {
-                'html': 'hint one',
-                'audio_translations': {}
-            }
-        })
-        state['interaction']['solution'] = solution
-        return state
 
     def setUp(self):
         """Before each individual test, create dummy user."""
